Flesh out name handling. Add MakeError.
Most functions should be returning a Go `error` interface, not the
GSSAPI `Status`; added `lib.MakeError()` to make it easy to get a nil or
an error type.  Need the lib there to have access to the parts-decoding
functionality.

A bunch of name handlers, now only missing the mech-list/name lookups.