Wrapper should force tests to run under Docker
We advertise this as the means to run under Docker, but on MacOS was
never doing so using the `run-heimdal.sh` wrapper; instead, ran
entrypoint on the local host.  If a non-system kinit is available, tests
would pass but leave litter in that cache.  If only the MacOS system
kinit is available, test would fail.

In the wrapper, set `CLIENT_IN_CONTAINER=yes` to force Docker, always.